--- src/fairseq2/utils/__init__.py
@@ -1,3 +1,0 @@
-from fairseq2.utils.read_and_pickle_wrap_iterator import (
-    read_and_pickle_wrap_iterator as read_and_pickle_wrap_iterator,
-)
